Updated package.json to use new SPDX license exp.
According to the docs on npmjs.com and the warning output when running
`npm install`, the license element in `package.json` was invalid. NPM
has updated their requirements to use the new SPDX license expression
syntax version 2.0, and the license expression that was being used is
now deprecated. I updated the license field to use the new valid
expression. For more information, please see:
https://docs.npmjs.com/files/package.json#license