Updated `.travis.yml` and added `.nvmrc`
Added `sudo: false` to `.travis.yml` to enable project to be built
on new Travis CI infrastructure. Moved `node_js` version in
`.travis.yml` to use 'stable' so that we are always running against the
latest version. Added a new `.nvmrc` so that when running tests locally
we are all using the same version of Node.